Refactor: Improve rollpe entry logic
